Fixes to support both firmware versions
It's now checked at the startup if the modem uses the older firmware (mid 2018) or the newer one.
Depending on this, two different implementations are injected
into the actual main/login method to keep the communal part as big as possible.
Any future firmware versions can then be added here.